requirements-39.txt
absl-py
accelerate
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anyio
asttokens
async-timeout
attrs
bayes
bayesian-optimization
beautifulsoup4
botorch
brotli
cartopy
catboost
cattrs
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
colorlog
contourpy
coremltools
cssselect2
cycler
datasets
decorator
dgl
dgllife
dill
distro
docx
einops
evaluate
exceptiongroup
executing
fastapi
ffmpy
filelock
fonttools
frozenlist
fsspec
future
geographiclib
gpytorch
gradio
gradio-client
graphviz
groovy
grpcio
h11
httpcore
httpx
huggingface-hub
hyperopt
idna
imbalanced-learn
imblearn
importlib-metadata
importlib-resources
iniconfig
ipython
ipython-pygments-lexers
jaxtyping
jedi
jinja2
jiter
joblib
jpype1
kiwisolver
lightgbm
linear-operator
llvmlite
lxml
mako
markdown
markdown-it-py
markupsafe
matplotlib
matplotlib-inline
mdurl
mlx
mordred
mordredcommunity
mpmath
multidict
multipledispatch
multiprocess
narwhals
networkx
numba
numpy
obspy
openai
opt-einsum
optuna
optuna-integration
orjson
outcome
packaging
padelpy
pandas
parso
patsy
pdfkit
pexpect
pillow
pip
plotly
pluggy
prompt-toolkit
propcache
protobuf
psutil
ptyprocess
pure-eval
py4j
pyaml
pyarrow
pycparser
pydantic
pydantic-core
pydub
pydyf
pygments
pyparsing
pyphen
pyproj
pyro-api
pyro-ppl
pyshp
pysocks
pytest
python-dateutil
python-dotenv
python-multipart
pytorch-tabnet
pytz
pyyaml
rdkit
regex
reportlab
requests
rf
rich
ruff
safehttpx
safetensors
scikit-learn
scipy
seaborn
selenium
semantic-version
sentencepiece
setuptools
shap
shapely
shellingham
six
sklearn-compat
slicer
sniffio
sortedcontainers
soupsieve
sqlalchemy
stack-data
starlette
statsmodels
sympy
tensorboard
tensorboard-data-server
threadpoolctl
tinycss2
tinyhtml5
tokenizers
tomli
tomlkit
torch
torch-geometric
torchaudio
torchdata
torchvision
tqdm
traitlets
transformers
trio
trio-websocket
typeguard
typer
typing-extensions
typing-inspection
tzdata
umap
urllib3
uvicorn
wcwidth
weasyprint
webdriver-manager
webencodings
websocket-client
websockets
werkzeug
wheel
wkhtmltopdf
wsproto
xgboost
xxhash
yarl
zipp
zopfli
